+++ Source/HelixToolkit.UWP/Controls/MouseHandlers/PanHandler.cs
@@ -44,11 +44,9 @@ namespace HelixToolkit.UWP
         public override void Delta(Vector2 e)
         {
             base.Delta(e);
-            if (Camera.CameraInternal.LookDirection.LengthSquared() < 1f && MouseDownNearestPoint3D.HasValue)
+            if (Camera.CameraInternal.LookDirection.LengthSquared() < 1e-5f)
             {
-                var look = Camera.CameraInternal.LookDirection.Normalized();
-                var v = MouseDownNearestPoint3D.Value - Camera.CameraInternal.Position;
-                Camera.CameraInternal.LookDirection = look * Math.Max(1, Vector3.Dot(v, look));
+                return;
             }
             var thisPoint3D = this.UnProject(e, this.panPoint3D, this.Camera.CameraInternal.LookDirection);
 
@@ -127,7 +125,7 @@ namespace HelixToolkit.UWP
         {
             base.Started(e);
             this.panPoint3D = this.Camera.CameraInternal.Target;
-            if (this.MouseDownNearestPoint3D != null)
+            if (this.MouseDownNearestPoint3D.HasValue)
             {
                 this.panPoint3D = this.MouseDownNearestPoint3D.Value;
             }