+++ Source/HelixToolkit.Wpf/Controls/CameraController/CameraController.cs
@@ -201,6 +201,20 @@ namespace HelixToolkit.Wpf
                 "RotateAroundMouseDownPoint", typeof(bool), typeof(CameraController), new UIPropertyMetadata(false));
 
         /// <summary>
+        /// Identifies the <see cref="FixedRotationPointEnabled"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty FixedRotationPointEnabledProperty =
+            DependencyProperty.Register(
+                "FixedRotationPointEnabled", typeof(bool), typeof(CameraController), new UIPropertyMetadata(false));
+
+        /// <summary>
+        /// Identifies the <see cref="FixedRotationPoint"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty FixedRotationPointProperty =
+            DependencyProperty.Register(
+                "FixedRotationPoint", typeof(Point3D), typeof(CameraController), new UIPropertyMetadata(default(Point3D)));
+
+        /// <summary>
         /// Identifies the <see cref="RotateCursor"/> dependency property.
         /// </summary>
         public static readonly DependencyProperty RotateCursorProperty = DependencyProperty.Register(
@@ -1058,6 +1072,40 @@ namespace HelixToolkit.Wpf
         }
 
         /// <summary>
+        /// Gets or sets a value indicating whether to rotate around a fixed point.
+        /// </summary>
+        /// <value> <c>true</c> if rotation around a fixed point is enabled; otherwise, <c>false</c> . </value>
+        public bool FixedRotationPointEnabled
+        {
+            get
+            {
+                return (bool)this.GetValue(FixedRotationPointEnabledProperty);
+            }
+
+            set
+            {
+                this.SetValue(FixedRotationPointEnabledProperty, value);
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets a value indicating the center of rotation.
+        /// </summary>
+        /// <value> <c>true</c> if rotation around a fixed point is enabled; otherwise, <c>false</c> . </value>
+        public Point3D FixedRotationPoint
+        {
+            get
+            {
+                return (Point3D)this.GetValue(FixedRotationPointProperty);
+            }
+
+            set
+            {
+                this.SetValue(FixedRotationPointProperty, value);
+            }
+        }
+
+        /// <summary>
         /// Gets or sets the rotate cursor.
         /// </summary>
         /// <value> The rotate cursor. </value>
@@ -1249,6 +1297,7 @@ namespace HelixToolkit.Wpf
             }
         }
 
+
         /// <summary>
         /// Efficiency option, lower values decrease computation time for camera interaction when
         /// RotateAroundMouseDownPoint or ZoomAroundMouseDownPoint is set to true in inspect mode.
@@ -1299,7 +1348,16 @@ namespace HelixToolkit.Wpf
                 return this.ActualCamera as PerspectiveCamera;
             }
         }
-
+        /// <summary>
+        /// Gets or sets a value indicating whether [limit FPS].
+        /// </summary>
+        /// <value>
+        ///   <c>true</c> if [limit FPS]; otherwise, <c>false</c>.
+        /// </value>
+        public bool LimitFPS { set; get; } = true;
+        #region Private Variables
+        private TimeSpan prevTime;
+        #endregion
         /// <summary>
         /// Adds the specified move force.
         /// </summary>
@@ -2140,6 +2198,11 @@ namespace HelixToolkit.Wpf
         /// </param>
         private void OnCompositionTargetRendering(object sender, RenderingEventArgs e)
         {
+            if (LimitFPS && prevTime == e.RenderingTime)
+            {
+                return;
+            }
+            prevTime = e.RenderingTime;
             var ticks = e.RenderingTime.Ticks;
             var time = 100e-9 * (ticks - this.lastTick);